Picture Classifier - Neural Networks Hopfield Algorithm
This was a project for my Neural Networks course. The program implements
a Hopfield net in order to classify distorted pictures. The program
takes in a set of pictures as training data. The testing data consists
of distorted pictures and it matches the distorted picture with the
training picture that has the most similarities. The hopfield algorithm
uses binary threshold nodes, so the input used is in ASCII using O's and
white space.